+++ bugowl/manage.py
@@ -7,7 +7,7 @@ import sys
 
 def main():
 	"""Run administrative tasks."""
-	os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'bugowl_websocket.settings')
+	os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'api.settings')
 	try:
 		from django.core.management import execute_from_command_line
 	except ImportError as exc:
